Introduced SecurityUtils to gather in a central place the handling of JEP 411 changes.
